Added ProxyKernelTests
