+++ src/Microsoft.DotNet.Interactive.Tests/CompositeKernelTests.cs
@@ -65,81 +65,6 @@ new [] {1,2,3}");
                   .ContainSingle<CommandHandled>(e => e.Command == fsharpCommand);
         }
 
-        [FactSkipLinux]
-        public async Task Handling_kernel_can_be_specified_using_kernel_name_as_a_directive_as_a_proxy_named_pipe()
-        {
-            var fSharpKernel = new FSharpKernel();
-            using var kernel = new CompositeKernel
-            {
-                fSharpKernel
-            }.UseProxyKernel();
-            kernel.DefaultKernelName = fSharpKernel.Name;
-
-            var pipeName = Guid.NewGuid().ToString();
-            using var cSharpKernel = new CSharpKernel();
-            Action doWait = () =>
-                Task.Run(() => NamedPipeKernelServer.WaitForConnection(cSharpKernel, pipeName));
-            doWait();
-
-            using var events = kernel.KernelEvents.ToSubscribedList();
-
-            var proxyCommand = new SubmitCode($"#!connect test {pipeName}");
-
-            await kernel.SendAsync(proxyCommand);
-
-            var proxyCommand2 = new SubmitCode(@"
-var x = 1 + 1;
-x", targetKernelName: "test");
-
-            await kernel.SendAsync(proxyCommand2);
-
-            events.Should()
-                  .ContainSingle<CommandHandled>(e => e.Command == proxyCommand);
-        }
-
-        [FactSkipLinux]
-        public async Task Handling_kernel_can_be_specified_using_kernel_name_as_a_directive_as_a_proxy_named_pipe2()
-        {
-            var fSharpKernel = new FSharpKernel();
-            using var kernel = new CompositeKernel
-            {
-                fSharpKernel
-            }.UseProxyKernel();
-            kernel.DefaultKernelName = fSharpKernel.Name;
-
-            var pipeName = Guid.NewGuid().ToString();
-            using var cSharpKernel = new CSharpKernel();
-            Action doWait = () =>
-                Task.Run(() => NamedPipeKernelServer.WaitForConnection(cSharpKernel, pipeName));
-            doWait();
-
-            using var events = kernel.KernelEvents.ToSubscribedList();
-
-            var proxyCommand = new SubmitCode($"#!connect test {pipeName}");
-
-            await kernel.SendAsync(proxyCommand);
-
-            var proxyCommand2 = new SubmitCode(@"
-#!test
-var x = 1 + 1;
-x");
-
-            await kernel.SendAsync(proxyCommand2);
-
-            var proxyCommand3 = new SubmitCode(@"
-#!test
-var y = x + x;
-y");
-
-            await kernel.SendAsync(proxyCommand3);
-
-            events.Should()
-                  .ContainSingle<CommandHandled>(e => e.Command == proxyCommand2);
-
-            events.Should()
-                  .ContainSingle<CommandHandled>(e => e.Command == proxyCommand3);
-        }
-
         [Fact]
         public async Task Handling_kernel_can_be_specified_using_kernel_alias_as_a_directive()
         {
